Fix conversion warnings, tweak mario input handlers
- Use !KeyPressed instead of KeyUp to trigger stop walking action
- Tweak mario jumping behaviour
- Put directx input version macro in right place
- Fix some type conversion warnings

+++ src/GameCuaTao/Direct2DGame/Game.cpp
@@ -79,7 +79,7 @@ void Game::Run()
 	while (IsRunning())
 	{
 		Tick();
-		auto deltaTime = gameTime.ElapsedGameTime.Milliseconds();
+		auto deltaTime = (int)gameTime.ElapsedGameTime.Milliseconds();
 
 		if (deltaTime >= GetTickPerFrame())
 		{
@@ -107,7 +107,7 @@ bool Game::IsRunning()
 	return true;
 }
 
-double Game::GetTickPerFrame()
+int Game::GetTickPerFrame()
 {
 	return 1000 / graphics->GetFramePerSecond();
 }
